// src/types/enums.ts
export type CountryStatus = 'not_visited' | 'visited' | 'planned';

export type Continent =
  | 'Africa'
  | 'Asia'
  | 'Europe'
  | 'North America'
  | 'South America'
  | 'Oceania'
  | 'Antarctica';

export type ThemeMode = 'light' | 'dark' | 'system';

export type MapProjection = 'geoEqualEarth' | 'geoMercator' | 'geoNaturalEarth1';

export type DateFormat = 'YYYY-MM-DD' | 'DD/MM/YYYY' | 'MM/DD/YYYY';